move metrics out of platform
