+++ .eslintignore
@@ -1,4 +1,5 @@
 coverage
+dist
 docs
 out
 node_modules
