+++ packages/dd-trace/src/platform/node/index.js
@@ -6,7 +6,6 @@ const now = require('./now')
 const env = require('./env')
 const validate = require('./validate')
 const request = require('./request')
-const metrics = require('./metrics')
 const plugins = require('../../plugins')
 const hostname = require('./hostname')
 const pkg = require('./pkg')
@@ -30,7 +29,6 @@ const platform = {
   service: () => process.env['AWS_LAMBDA_FUNCTION_NAME'] || pkg.name,
   appVersion: () => pkg.version,
   request,
-  metrics,
   plugins,
   startupLog,
   hostname,
