+++ packages/dd-trace/test/setup/core.js
@@ -6,7 +6,7 @@ const sinonChai = require('sinon-chai')
 const os = require('os')
 const proxyquire = require('../proxyquire')
 const semver = require('semver')
-const platform = require('../../src/platform')
+const metrics = require('../../src/metrics')
 const AsyncHooksScope = require('../../src/scope/async_hooks')
 const AsyncLocalStorageScope = require('../../src/scope/async_local_storage')
 const AsyncResourceScope = require('../../src/scope/async_resource')
@@ -37,7 +37,7 @@ global.withVersions = withVersions
 
 afterEach(() => {
   agent.reset()
-  platform.metrics().stop()
+  metrics.stop()
 })
 
 function wrapIt (whichScope = defaultScope) {
